sys: bare-metal CPU usage parsing (fix)
* fix /proc/stat parsing: include specific enumerated fields
  - comments inline
* add temporary minimum wall-clock interval guard to reject
  millisecond-range resampling noise
  - TODO inline
* switch CPU load reporting to integer percentages
  - update all callers

+++ cmn/load/npa.go
@@ -106,7 +106,7 @@ func CPU() Load {
 	switch {
 	case extreme:
 		return Critical
-	case load >= float64(high):
+	case load >= high:
 		return High
 	default:
 		return Low
